Refactor mute handling in test scripts and update workflow paths
- Replaced the deprecated `resolve_muted_ya_path.py` with `mute_check.py` for resolving mute paths in the GitHub Actions workflow.
- Updated import statements across various scripts to reflect the new structure under the `mute` directory.
- Enhanced the `update_muted_ya.yml` workflow to utilize the new mute management scripts, improving clarity and maintainability.
